+++ CHANGELOG.md
@@ -53,6 +53,7 @@ and try to associate the created connection with this project automatically.
 * [#707](https://github.com/clojure-emacs/cider/issues/707): Better support clojure.test/with-test.
 * Fix namespace navigation in the namespace browser.
 * [#1565](https://github.com/clojure-emacs/cider/issues/1565): Fix font-locking in apropos buffers.
+* [#1570](https://github.com/clojure-emacs/cider/issues/1570): Handle properly rest params in eldoc.
 
 ## 0.10.2 / 2016-01-27